fix: fix warnings
