+++ src/hooks/useTheme.tsx
@@ -1,5 +1,5 @@
 import { useContext } from 'react';
-import { ThemeContext } from '../context/ThemeContext';
+import { ThemeContext } from '../context/themeContextCreation';
 
 export const useTheme = () => {
   const context = useContext(ThemeContext);
